"check_style_score -> 0 of 5 passed. style score: 0/1 "

+++ cspp2-assignments/m2/Assignment-3/Solution.java
@@ -6,15 +6,15 @@ public class Solution {
 	*/
 	public static void main(String[] args) {
 		Scanner s=new Scanner(System.in);      
-        int base = s.nextInt();
-        int exponent = s.nextInt();
-        int result=power(base,exponent);
+        double base = s.nextInt();
+        double exponent = s.nextInt();
+        double result=power(base,exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int base, int exponent){
+	public static double power(double base, double exponent){
 		if (exponent!=0){
 			return (base * power(base, exponent-1));
 		}
